requirements.txt
boto3
langchain
langchain-community
mysqlclient
python-dotenv